Refactor environment_variables directory structure
