+++ lib/ey-core/version.rb
@@ -1,5 +1,5 @@
 module Ey
   module Core
-    VERSION = "3.6.4"
+    VERSION = "3.6.5"
   end
 end
